pyproject.toml
[tool.poetry]
name = "earningspredict-data-pipeline"
version = "0.1.0"
description = "Data pipeline for gathering information for EarningsPredict analysis."
authors = ["peppapig450 <nbgenius1@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
alpha-vantage = "^2.3.1"
fmpsdk = "^20240330.0"
finnhub-python = "^2.4.19"
pandas = "^2.2.2"
aiohttp = "^3.9.5"
tqdm = "^4.66.4"
pyarrow = "^16.1.0"
pytest-asyncio = "^0.23.7"
aiofiles = "^23.2.1"
pydantic = "^2.7.1"
more-itertools = "^10.2.0"
pytest = "^7.0"
pytest-mock = "^3.14.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
